refactor(@types): organiza pasta de tipos em subdiretorios por contexto
